refactor the name of class to follow python naming rules put __initi__ at the beginning of the class remove some variable into internal variable to simply functions and reduce passing variable everywhere.
